+++ internal/impl/grpc/common.go
@@ -1,4 +1,4 @@
-package grpc_client
+package grpc
 
 const (
 	rpcTypeUnary        = "unary"
